+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Initial Implementation
+- Initial Implementation ([#13])
+- Enabled monitoring using the Prometheus emitter extension. ([#27])
 
 [#13]: https://github.com/stackabletech/druid-operator/pull/13
+[#27]: https://github.com/stackabletech/druid-operator/pull/27